Move VQRSolver and VQRSolution to new solvers.base
