+++ src/vqr/solvers/__init__.py
@@ -1,5 +1,6 @@
 from vqr.solvers.cvx import CVXVQRSolver
 from vqr.solvers.pot import POTVQESolver
+from vqr.solvers.base import VQRSolver, VQRSolution
 from vqr.solvers.regularized_lse import (
     RegularizedDualVQRSolver,
     MLPRegularizedDualVQRSolver,
