+++ src/vqr/solvers/regularized_lse.py
@@ -15,9 +15,9 @@ from tqdm.auto import tqdm
 from torch.optim import Optimizer
 from torch.optim.lr_scheduler import ReduceLROnPlateau
 
-from vqr import VQRSolver, VQRSolution
 from vqr.vqr import vector_quantile_levels
 from vqr.models import MLP
+from vqr.solvers.base import VQRSolver, VQRSolution
 
 _LOG = logging.getLogger(__name__)
